# cpr-backend/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger CPR - OpenAPI 3.0
  description: |-
    Đây là API của website CPR
  version: 1.0.1
servers:
...
paths:
...
  /api/user/all:
    get:
      tags:
        - User
      summary: Get all users
      description: Return a list of users
      operationId: getAllUsers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
  /api/user:
    post:
      tags:
        - User
      summary: Create user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              required:
                - username
                - email
                - phoneNumber
                - address
              type: object
              properties:
                username:
                  type: string
                  example: Tiến Tài
                email:
                  type: string
                  example: tientai@example.com
                password:
                  type: string
                  example: ""
                phoneNumber:
                  type: string
                  example: 0907778777
                address:
                  type: string
                  example: Thành phố HCM
      responses:
        "201":
          description: Create user success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
    get:
      tags:
        - User
      summary: Get all users with searchTerm and paging
      description: Return a list of users
      operationId: getAllUsersWithPaging
      parameters:
        - name: searchTerm
          in: query
          description: The search term for users
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: The page number for users
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: The limit number for users
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
...
  /api/user/{username}:
    get:
      tags:
        - User
      summary: Get user by username
      description: Returns an user
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: "The username of the user"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    put:
      tags:
        - User
      summary: Update user by username
      description: Update an existing user by username
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: "The username of the user"
          required: true
          schema:
            type: string
      requestBody:
        description: update user object
        content:
          application/json:
            schema:
              required:
                - username
                - email
                - phoneNumber
                - address
              type: object
              properties:
                username:
                  type: string
                  example: "Tiến Tài"
                email:
                  type: string
                  example: tientai@example.com
                password:
                  type: string
                  example: ""
                phoneNumber:
                  type: string
                  example: 0907778777
                address:
                  type: string
                  example: Thành phố HCM
                status:
                  type: string
                  example: "Active"
                  enum:
                    - Active
                    - Inactive
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    delete:
      tags:
        - User
      summary: Delete user by username
      description: Delete an existing user by username
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: "The username of the user"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "204":
          description: Delete user success
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
...
components:
  schemas:
...
    User:
      required:
        - username
        - email
        - phoneNumber
        - address
      type: object
      properties:
        username:
          type: string
          example: "Tiến Tài"
        email:
          type: string
          example: tientai@example.com
        password:
          type: string
          example: "Pa$$w0rd"
        phoneNumber:
          type: string
          example: "12345"
        address:
          type: string
          example: "Thành phố HCM"
        role:
          type: string
          example: "user"
          enum:
            - admin
            - user
        status:
          type: string
          description: User Status
          example: Active
          enum:
            - Active
            - Inactive
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header